pokedex_backend/requirements.txt
fastapi>=0.68.0
sqlmodel>=0.0.8
pillow>=9.0.0
uvicorn>=0.15.0
aiofiles>=0.7.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
